@@ -3,8 +3,6 @@ import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat/src/core/api/channel_api.dart';
-import 'package:stream_chat/src/core/models/channel_model.dart';
-import 'package:stream_chat/src/core/models/channel_state.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:test/test.dart';
 
@@ -177,14 +175,14 @@ void main() {
 
   test('markAllRead', () async {
     const path = '/channels/read';
-    when(() => client.post(path)).thenAnswer(
+    when(() => client.post(path, data: {})).thenAnswer(
         (_) async => successResponse(path, data: <String, dynamic>{}));
 
     final res = await channelApi.markAllRead();
 
     expect(res, isNotNull);
 
-    verify(() => client.post(path)).called(1);
+    verify(() => client.post(path, data: {})).called(1);
     verifyNoMoreInteractions(client);
   });
 
