@@ -2,9 +2,7 @@ import 'dart:convert';
 
 import 'package:dio/dio.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:stream_chat/src/core/api/requests.dart';
 import 'package:stream_chat/src/core/api/user_api.dart';
-import 'package:stream_chat/src/core/models/filter.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:test/test.dart';
 
@@ -82,6 +80,39 @@ void main() {
     expect(res.users.length, updatedUsers.length);
 
     verify(() => client.post(path, data: any(named: 'data'))).called(1);
+    verifyNoMoreInteractions(client);
+  });
+
+  test('partialUpdateUsers', () async {
+    const user = PartialUpdateUserRequest(
+      id: 'test-user-id',
+      set: {'color': 'yellow'},
+    );
+
+    const path = '/users';
+
+    final updatedUser = {user.id: User(id: user.id, extraData: user.set!)};
+
+    when(() => client.patch(path, data: {
+          'users': [user],
+        })).thenAnswer(
+      (_) async => successResponse(
+        path,
+        data: {
+          'users':
+              updatedUser.map((key, value) => MapEntry(key, value.toJson()))
+        },
+      ),
+    );
+
+    final res = await userApi.partialUpdateUsers([user]);
+
+    expect(res, isNotNull);
+    expect(res.users.length, updatedUser.length);
+
+    verify(() => client.patch(path, data: {
+          'users': [user]
+        })).called(1);
     verifyNoMoreInteractions(client);
   });
 }
