@@ -4,7 +4,6 @@ import 'dart:math';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:collection/collection.dart';
 import 'package:file_picker/file_picker.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
@@ -16,16 +15,12 @@ import 'package:stream_chat_flutter/src/emoji/emoji.dart';
 import 'package:stream_chat_flutter/src/emoji_overlay.dart';
 import 'package:stream_chat_flutter/src/extension.dart';
 import 'package:stream_chat_flutter/src/media_list_view.dart';
-import 'package:stream_chat_flutter/src/message_list_view.dart';
 import 'package:stream_chat_flutter/src/multi_overlay.dart';
 import 'package:stream_chat_flutter/src/quoted_message_widget.dart';
-import 'package:stream_chat_flutter/src/stream_chat_theme.dart';
-import 'package:stream_chat_flutter/src/stream_svg_icon.dart';
 import 'package:stream_chat_flutter/src/user_mentions_overlay.dart';
 import 'package:stream_chat_flutter/src/video_service.dart';
 import 'package:stream_chat_flutter/src/video_thumbnail_image.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 import 'package:video_compress/video_compress.dart';
 
 export 'package:video_compress/video_compress.dart' show VideoQuality;
@@ -384,6 +379,7 @@ class MessageInputState extends State<MessageInput> {
 
   final _imagePicker = ImagePicker();
   late final _focusNode = widget.focusNode ?? FocusNode();
+  late final _isInternalFocusNode = widget.focusNode == null;
   bool _inputEnabled = true;
   bool _commandEnabled = false;
   bool _showCommandsOverlay = false;
@@ -1197,7 +1193,7 @@ class MessageInputState extends State<MessageInput> {
           quality: widget.compressedVideoQuality,
         );
 
-        if (mediaInfo!.filesize! > maxAttachmentSize) {
+        if (mediaInfo == null || mediaInfo.filesize! > maxAttachmentSize) {
           _showErrorAlert(
             context.translations.fileTooLargeAfterCompressionError(
               maxAttachmentSize / (1024 * 1024),
@@ -1257,30 +1253,36 @@ class MessageInputState extends State<MessageInput> {
       };
     }
 
-    return UserMentionsOverlay(
-      query: query,
-      mentionAllAppUsers: widget.mentionAllAppUsers,
-      client: StreamChat.of(context).client,
-      channel: channel,
-      size: Size(renderObject.size.width - 16, 400),
-      mentionsTileBuilder: tileBuilder,
-      onMentionUserTap: (user) {
-        _mentionedUsers.add(user);
-        splits[splits.length - 1] = user.name;
-        final rejoin = splits.join('@');
+    return LayoutBuilder(
+      builder: (context, snapshot) => UserMentionsOverlay(
+        query: query,
+        mentionAllAppUsers: widget.mentionAllAppUsers,
+        client: StreamChat.of(context).client,
+        channel: channel,
+        size: Size(
+          renderObject.size.width - 16,
+          min(400, (snapshot.maxHeight - renderObject.size.height - 16).abs()),
+        ),
+        mentionsTileBuilder: tileBuilder,
+        onMentionUserTap: (user) {
+          _mentionedUsers.add(user);
+          splits[splits.length - 1] = user.name;
+          final rejoin = splits.join('@');
 
-        textEditingController.value = TextEditingValue(
-          text: rejoin +
-              textEditingController.text.substring(
-                textEditingController.selection.start,
-              ),
-          selection: TextSelection.collapsed(
-            offset: rejoin.length,
-          ),
-        );
-        _onChangedDebounced.cancel();
-        setState(() => _showMentionsOverlay = false);
-      },
+          textEditingController.value = TextEditingValue(
+            text: rejoin +
+                textEditingController.text.substring(
+                  textEditingController.selection.start,
+                ),
+            selection: TextSelection.collapsed(
+              offset: rejoin.length,
+            ),
+          );
+          _onChangedDebounced.cancel();
+
+          setState(() => _showMentionsOverlay = false);
+        },
+      ),
     );
   }
 
@@ -1924,6 +1926,8 @@ class MessageInputState extends State<MessageInput> {
 
     _mentionedUsers.clear();
 
+    message = _replaceUserNameWithId(message);
+
     try {
       Future sendingFuture;
       if (widget.editMessage == null ||
@@ -2031,6 +2035,7 @@ class MessageInputState extends State<MessageInput> {
   void dispose() {
     textEditingController.dispose();
     _focusNode.removeListener(_focusNodeListener);
+    if (_isInternalFocusNode) _focusNode.dispose();
     _stopSlowMode();
     _onChangedDebounced.cancel();
     super.dispose();
@@ -2076,12 +2081,12 @@ class _PickerWidget extends StatefulWidget {
 }
 
 class _PickerWidgetState extends State<_PickerWidget> {
-  Future<bool>? requestPermission;
+  Future<PermissionState>? requestPermission;
 
   @override
   void initState() {
     super.initState();
-    requestPermission = PhotoManager.requestPermission();
+    requestPermission = PhotoManager.requestPermissionExtend();
   }
 
   @override
@@ -2089,14 +2094,15 @@ class _PickerWidgetState extends State<_PickerWidget> {
     if (widget.filePickerIndex != 0) {
       return const Offstage();
     }
-    return FutureBuilder<bool>(
+    return FutureBuilder<PermissionState>(
       future: requestPermission,
       builder: (context, snapshot) {
         if (!snapshot.hasData) {
           return const Offstage();
         }
 
-        if (snapshot.data!) {
+        if ([PermissionState.authorized, PermissionState.limited]
+            .contains(snapshot.data)) {
           if (widget.containsFile) {
             return GestureDetector(
               onTap: () {
@@ -2188,4 +2194,22 @@ class _CountdownButton extends StatelessWidget {
           ),
         ),
       );
+}
+
+Message _replaceUserNameWithId(Message message) {
+  final mentionedUsers = message.mentionedUsers;
+  if (mentionedUsers.isEmpty) return message;
+
+  var messageTextToSend = message.text;
+  if (messageTextToSend == null) return message;
+
+  for (final user in mentionedUsers.toSet()) {
+    final userName = user.name;
+    messageTextToSend = messageTextToSend!.replaceAll(
+      '@$userName',
+      '@${user.id}',
+    );
+  }
+
+  return message.copyWith(text: messageTextToSend);
 }
