@@ -1,11 +1,11 @@
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:stream_chat_flutter_core/src/channel_list_core.dart';
 import 'package:stream_chat_flutter_core/src/stream_chat_core.dart';
+import 'package:stream_chat_flutter_core/src/stream_controller_extension.dart';
 
 /// Widget dedicated to the management of a channel list with pagination
 /// [ChannelsBloc] is used together with [ChannelListCore] to manage a list of
@@ -116,7 +116,7 @@ class ChannelsBlocState extends State<ChannelsBloc>
     }
 
     if (_channelsController.hasValue) {
-      _queryChannelsLoadingController.add(true);
+      _queryChannelsLoadingController.safeAdd(true);
     }
 
     try {
@@ -135,14 +135,14 @@ class ChannelsBlocState extends State<ChannelsBloc>
       )) {
         newChannels = channels;
         if (clear) {
-          _channelsController.add(channels);
+          _channelsController.safeAdd(channels);
         } else {
           final temp = oldChannels + channels;
-          _channelsController.add(temp);
+          _channelsController.safeAdd(temp);
         }
         if (_channelsController.hasValue &&
             _queryChannelsLoadingController.value) {
-          _queryChannelsLoadingController.sink.add(false);
+          _queryChannelsLoadingController.safeAdd(false);
         }
       }
       if (newChannels.isEmpty || newChannels.length < paginationParams.limit) {
@@ -150,11 +150,11 @@ class ChannelsBlocState extends State<ChannelsBloc>
       }
     } catch (e, stk) {
       // reset loading controller
-      _queryChannelsLoadingController.sink.add(false);
+      _queryChannelsLoadingController.safeAdd(false);
       if (_channelsController.hasValue) {
-        _queryChannelsLoadingController.addError(e, stk);
+        _queryChannelsLoadingController.safeAddError(e, stk);
       } else {
-        _channelsController.addError(e, stk);
+        _channelsController.safeAddError(e, stk);
       }
     }
   }
@@ -175,6 +175,9 @@ class ChannelsBlocState extends State<ChannelsBloc>
           'time_updated',
         )
             .listen((e) {
+          if (e.message?.parentId != null && e.message?.showInChannel != true) {
+            return;
+          }
           final newChannels = List<Channel>.from(channels ?? []);
           final index = newChannels.indexWhere((c) => c.cid == e.cid);
           if (index != -1) {
@@ -198,7 +201,7 @@ class ChannelsBlocState extends State<ChannelsBloc>
           if (widget.channelsComparator != null) {
             newChannels.sort(widget.channelsComparator);
           }
-          _channelsController.add(newChannels);
+          _channelsController.safeAdd(newChannels);
         }));
       }
 
@@ -210,7 +213,7 @@ class ChannelsBlocState extends State<ChannelsBloc>
           if (channelIndex > -1) {
             final channel = newChannels.removeAt(channelIndex);
             _hiddenChannels.add(channel);
-            _channelsController.add(newChannels);
+            _channelsController.safeAdd(newChannels);
           }
         }))
         ..add(client
@@ -220,7 +223,7 @@ class ChannelsBlocState extends State<ChannelsBloc>
         )
             .listen((e) {
           final channel = e.channel;
-          _channelsController.add(List.from(
+          _channelsController.safeAdd(List.from(
             (channels ?? [])..removeWhere((c) => c.cid == channel?.cid),
           ));
         }));
